1、升级 spring boot 2.3.2 mybatis-plus 3.2.0 等 2、Angular 第一个请求后台 3、因为 mybatis-plus 升级有较大的改动，controller 和 service 层也有较大的调整 4、还有其他大的调整

+++ app_web/src/main/java/com/app/web/vo/PageVo.java
@@ -1,6 +1,6 @@
 package com.app.web.vo;
 
-import com.baomidou.mybatisplus.plugins.Page;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import lombok.Data;
 
 import java.util.List;
@@ -20,10 +20,10 @@ public class PageVo<T> {
 
     public static <M> PageVo<M> parseMybatisPlusPage(Page<M> page){
         PageVo<M> pageVo = new PageVo<>();
-        pageVo.setPage(page.getCurrent());
-        pageVo.setPerPage(page.getSize());
-        pageVo.setTotalPage(page.getPages());
-        pageVo.setTotalSize(page.getTotal());
+        pageVo.setPage((int) page.getCurrent());
+        pageVo.setPerPage((int)  page.getSize());
+        pageVo.setTotalPage((int)  page.getPages());
+        pageVo.setTotalSize((int)  page.getTotal());
         pageVo.setData(page.getRecords());
         return pageVo;
     }
